Remove MarshalJSON and UnmarshalJSON methods
This commit removes the MarshalJSON and UnmarshalJSON methods from the
generated code. This does not break JSON support since the generated
code implements the encoding.TextMarshaler and encoding.TextUnmarshaler
methods, which the json package will use.

There are two reasons for this change:
1. Some rough testing shows that encoding a TextMarshaler to JSON is
   roughly 25% fast than encoding a json.Marshaler. This is likely due
   to output of MarshalText() not being compacted by the json package
   (since it is assumed to be a string whereas the output of MarshalJSON
   can be anything). The output of MarshalText only has to pass through
   the stringBytes() function of the json package.
2. Not implementing MarshalJSON and UnmarshalJSON allows us to support
   enum strings that are not valid JSON (possible with the -linecomment
   option) without having to add a tremendous amount of near duplicate
   code.

TLDR: this makes JSON encoding faster with not cost to decoding and
allows us to support arbitrary user supplied strings (via the
-linecomment option).