go-enum: add support for generating sql.Scanner and database.Valuer
This commit adds support via the `-sql` flag to generate methods that
implement the sql.Scanner and sql/driver.Valuer interfaces using the
string value of the enum.

TODO: add support for integer based database enums (this is only really
needed if the type is not defined as an ENUM in the database).